Move TypedDict-Kwargs to dedicated file
